Imported a self drawn image of the health bar to make it more fitting with the theme of the game.
I also tried to make the homing missile but that didn't quite work out unfortunately. There's still some work there to be done.

Started working on the ammo count for the hud so I created a new WBP for it but since the tutorial I followed had a different way of spawning the projectiles and equipped weapons I tried to mold it to our needs but this didn't work out and now the code doesn't work